bandwidth jugaad: pure-Go sqlite (modernc, CGO_ENABLED=0) + Back4app/Serv00 deploy files + optional Storj egress cap (default OFF)
- sqlite driver: mattn (CGO) -> modernc.org/sqlite (pure-Go) — main.go + upstash.go driver name 'sqlite'
- sqlstore.New 'sqlite3' -> 'sqlite' (unknown-driver crash fix, CGO-free binary ke saath)
- gold-cmds/system.go: FreeBSD Statfs uint64 cast fix (Serv00 build)
- storj.go: GOLDMD_STORJ_MAX_UPLOAD_KB env-gated proto egress cap — DEFAULT OFF (purana behaviour 50MiB tak sab upload)
- gold-md-cgofree: Linux amd64 static binary (CGO_ENABLED=0, -s -w, 22MB)
- gold-md-freebsd: FreeBSD amd64 static binary (Serv00)
- Dockerfile.back4app: debian slim + ffmpeg/python3-pil/poppler/libreoffice + prebuilt binary (PORT 2081)
- deploy_serv00.sh: binexec + port + www proxy + cron watchdog
- BANDWIDTH_JUGAAD_GUIDE.md: Roman-Urdu migration guide (Render 5GB -> Back4app 100GB)
- vendor: modernc.org/sqlite + deps added, mattn/go-sqlite3 removed, whatsmeow patches restored (clientpayload.go waVersion 1045649367 + broadcast.go LID 479 fix)

Bot behaviour bilkul pehle jaisa — sab optional env-gated hai.

+++ gold-cmds/system.go
@@ -360,8 +360,8 @@ func readDiskInfo(path string) diskInfo {
 		return diskInfo{"N/A", "N/A", "N/A", "N/A", 0, 0}
 	}
 	blockSize := uint64(stat.Bsize)
-	total := stat.Blocks * blockSize
-	free := stat.Bavail * blockSize
+	total := uint64(stat.Blocks) * blockSize
+	free := uint64(stat.Bavail) * blockSize
 	used := uint64(0)
 	if total > free {
 		used = total - free
